Remove dots around "jbovlaste" in README
The rule about dots surrounding names — dotside or no — refers to _cmevla_, not to "anything that follows `la`". And `jbovlaste` is a brivla, not a cmevla.

(It's not strictly _wrong_ to pause here, but it betrays some confusion about the point of the dots, and propagates the wrong idea about when dots are necessary.)

Compare to [section 6.2 in CLL](https://lojban.org/publications/cll/cll_v1.1_xhtml-section-chunks/section-basic-descriptors.html) which talks about `la cribe`, not `la .cribe.`.